backend/requirements.txt
fastapi~=0.110.0
uvicorn[standard]~=0.38.0
pydantic~=2.0
pydantic-settings~=2.0
sqlmodel~=0.0.31
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.9
langchain~=0.1.0
sentence-transformers~=2.5.1
pypdf~=4.1.0
openai~=1.13.3
psycopg2-binary~=2.9.9
pgvector~=0.2.5
alembic~=1.13.1
pyotp==2.9.0
qrcode==7.4.2
pillow==10.2.0
exifread==3.0.0
reportlab==4.0.9
pandas~=2.2.1
openpyxl~=3.1.2
google-generativeai~=0.4.1
thefuzz~=0.22.1
python-Levenshtein~=0.25.0
pytest~=8.1.1
pytest-asyncio~=0.23.5
anyio~=4.3.0
# Background Task Processing
celery[redis]~=5.3.6
redis~=5.0.1
flower~=2.0.1
psutil~=5.9.8
